We're going 32 bit

+++ headers/helpers.h
@@ -1,11 +1,11 @@
 #ifndef HELPERS_H
 #define HELPERS_H
 
-double calculateDistanceArray(double *p, int start, double *ref, uint dims);
+float calculateDistanceArray(float *p, int start, float *ref, uint dims);
 
-uint partition(double *arr, uint low, uint high);
-double kthSmallest(double *array, uint start, uint end, uint k);
-double quickselect(double *distances, uint end);
+uint partition(float *arr, uint low, uint high);
+float kthSmallest(float *array, uint start, uint end, uint k);
+float quickselect(float *distances, uint end);
 
 
 #endif
